code/gguf-model-api/pyproject.toml
[project]
name = "gguf-model-api"
version = "0.1.0"
description = "A simple FastAPI app serving a gguf model trained on function calling"
authors = [
    {name = "Monade",email = "team@monade.io"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = "^3.12.7"
dependencies = [
    "transformers (>=4.53.1,<5.0.0)",
    "fastapi[standard] (>=0.115.14,<0.116.0)",
    "huggingface-hub (>=0.33.2,<0.34.0)",
    "llama-cpp-python (>=0.3.10,<0.4.0)"
]

[tool.poetry]
package-mode = false

[tool.poetry.scripts]
gguf-model-api = "main:main"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
